src/main/java/com/wechat/sell/service/impl/ProductInfoServiceImpl.java
package com.wechat.sell.service.impl;

import com.wechat.sell.dao.ProductInfoDao;
import com.wechat.sell.dto.CartDTO;
import com.wechat.sell.entity.ProductInfo;
import com.wechat.sell.enums.ProductStatusEnum;
import com.wechat.sell.enums.ResultEnum;
import com.wechat.sell.exception.SellException;
import com.wechat.sell.service.ProductInfoService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import java.util.List;

@Service
public class ProductInfoServiceImpl implements ProductInfoService {

    @Autowired
    private ProductInfoDao productInforepository;

    @Override
    public ProductInfo findOne(String productId) {
        return productInforepository.findById(productId).get();
}

    @Override
    public Page<ProductInfo> findAll(Pageable pageable) {
        return productInforepository.findAll(pageable);
    }

    @Override
    public List<ProductInfo> findByProductStatus() {
        return productInforepository.findByProductStatus(ProductStatusEnum.UP.getCode());
    }

    @Override
    public void save(ProductInfo productInfo) {
        productInforepository.save(productInfo);
    }

    @Override
    @Transactional
    public void increaseStock(List<CartDTO> cartDTOList) {
        for (CartDTO cartDTO: cartDTOList) {
            ProductInfo productInfo = productInforepository.findById(cartDTO.getProductId()).get();
            if (productInfo == null) {
                throw new SellException(ResultEnum.PRODUCT_NOT_EXIST);
            }
            Integer result = productInfo.getProductStock() + cartDTO.getProductQuantity();
            productInfo.setProductStock(result);

            productInforepository.save(productInfo);
        }

    }

    @Override
    @Transactional
    public void decreaseStock(List<CartDTO> cartDTOList) {
        for (CartDTO cartDTO: cartDTOList) {
            ProductInfo productInfo = productInforepository.findById(cartDTO.getProductId()).get();
            if (productInfo == null) {
                throw new SellException(ResultEnum.PRODUCT_NOT_EXIST);
            }

            Integer result = productInfo.getProductStock() - cartDTO.getProductQuantity();
            if (result < 0) {
                throw new SellException(ResultEnum.PRODUCT_STOCK_ERROR);
            }

            productInfo.setProductStock(result);

            productInforepository.save(productInfo);
        }
    }

    @Override
    public ProductInfo onSale(String productId) {
       /* ProductInfo productInfo = productInforepository.findById(productId).get();
        if (productInfo == null) {
            throw new SellException(ResultEnum.PRODUCT_NOT_EXIST);
        }
        if (productInfo.getProductStatusEnum() == ProductStatusEnum.UP) {
            throw new SellException(ResultEnum.PRODUCT_STATUS_ERROR);
        }

        //更新
        productInfo.setProductStatus(ProductStatusEnum.UP.getCode());
        return productInforepository.save(productInfo);*/
       return null;
    }

    @Override
    public ProductInfo offSale(String productId) {
       /* ProductInfo productInfo = productInforepository.findById(productId).get();
        if (productInfo == null) {
            throw new SellException(ResultEnum.PRODUCT_NOT_EXIST);
        }
        if (productInfo.getProductStatusEnum() == ProductStatusEnum.DOWN) {
            throw new SellException(ResultEnum.PRODUCT_STATUS_ERROR);
        }

        //更新
        productInfo.setProductStatus(ProductStatusEnum.DOWN.getCode());
        return productInforepository.save(productInfo);*/
        return null;
    }
}
